add foreign classes

+++ src/main/kotlin/com/s63d/erinvoiceservice/domain/rest/InternResponse.kt
@@ -1,3 +1,3 @@
 package com.s63d.erinvoiceservice.domain.rest
 
-data class InternResponse(val id: String, val distance: Double, val price: Double, val rate: Double, val label: String, val dest: String)
+data class InternResponse(val id: String, val distance: Long, val price: Double, val rate: Double, val label: String, val dest: String)
